+++ CHANGELOG.md
@@ -1,14 +1,20 @@
-# latest
+# 0.2.0.0
 
 * Add `map'` & `unzipWith`
 * Make `zip` & `unzip` more strict
-* Make various functions strict in the index
+* Make all functions strict in the index
 * Make the `FoldableWithIndex`, `FunctorWithIndex`, `TraversableWithIndex` instances more efficient
+* Implement `ifoldMap` in the `FoldableWithIndex` instance
+* Implement `fail` in the `Monad` instance
+* Add `Shift`, `Tree` and unidirectional pattern synonyms for `Vector` and `Tree` to `Data.RRBVector.Internal.Debug`
+
+* Test typeclass laws
 
 # 0.1.1.0 - August 2021
 
 * Add `replicate`
 * Various optimizations
+
 * Switch to the [`tasty` framework](https://hackage.haskell.org/package/tasty) for benchmarks and tests
 * Extend the test suite
   - Add properties for more functions and some typeclass instances